Added vscode to gitignore, splited unit and integration test file and changed id_board read method
