+++ tests/test_app.py
@@ -4,19 +4,7 @@ from todo_app.app import create_app
 from dotenv import find_dotenv,load_dotenv
 import datetime as dt
 import pytest
-
-@pytest.fixture
-def client():
-    # Use our test integration config instead of the "real" version
-    file_path = find_dotenv('.env.test') 
-    load_dotenv(file_path, override=True)
-
-    # Create the new app.
-    test_app = create_app()
-
-    # Use the app to create a test_client that can be used in our tests.
-    with test_app.test_client() as client:
-        yield client
+import json
 
 class TestUnit:
     @staticmethod
@@ -92,9 +80,3 @@ class TestUnit:
                 bOK = False
                 break
         assert bOK == True
-
-class TestIntegration:
-    @staticmethod
-    def test_index_page(client):
-        response = client.get('/')
-        assert response.status_code == 200
